# validation/results/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "MolDesign API",
  "summary": "Plataforma de dise\u00f1o molecular asistido con pipeline cient\u00edfico reproducible.",
  "description": "MolDesign expone un pipeline cient\u00edfico basado en RDKit, AutoDock Vina y scoring expl\u00edcito. La IA interpreta resultados ya calculados y no genera m\u00e9tricas qu\u00edmicas por s\u00ed misma.",
  "version": "1.0.0-alpha.2"
 },
 "paths": {
...
  "/evaluation/cohorts/preflight": {
   "post": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Cohortes"
    ],
    "summary": "Comprobaci\u00f3n previa de una cohorte (no ejecuta docking ni la persiste)",
    "description": "Analiza y normaliza una cohorte sin ejecutarla y sin guardarla.\n\nSuperar esta comprobaci\u00f3n **no predice uni\u00f3n ni calidad farmacol\u00f3gica**: no\nse ha calculado nada todav\u00eda. S\u00f3lo dice que estas entradas se pueden leer y\nque esta configuraci\u00f3n se puede aplicar a todas por igual.",
    "operationId": "cohort_preflight_evaluation_cohorts_preflight_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_cohort_preflight_evaluation_cohorts_preflight_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CohortPreflightResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/evaluation/cohorts": {
   "post": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Cohortes"
    ],
    "summary": "Congelar una cohorte comprobada (no ejecuta nada)",
    "description": "Acepta una comprobaci\u00f3n previa y la congela como cohorte durable.\n\nLo que queda congelado: la definici\u00f3n normalizada, el veredicto entero \u2014con\nsus filas inv\u00e1lidas y duplicadas\u2014, el archivo original y su SHA-256. Lo que\nNO ocurre: ning\u00fan c\u00e1lculo cient\u00edfico. `ready` sigue significando lo mismo\nque en la comprobaci\u00f3n previa \u2014entrada ejecutable\u2014 y no hay todav\u00eda ninguna\nevidencia de docking.",
    "operationId": "create_cohort_evaluation_cohorts_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_cohort_evaluation_cohorts_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CohortRecord"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Cohortes"
    ],
    "summary": "Cohortes congeladas (resumen: sin filas y sin archivo)",
    "description": "Las cohortes de quien pregunta, de la m\u00e1s reciente a la m\u00e1s antigua.\n\nSin filas y sin bytes: este listado contesta \u00ab\u00bfqu\u00e9 cohortes tengo?\u00bb, y\ncontestarlo con 500 filas por cohorte lo convertir\u00eda en una descarga.",
    "operationId": "list_cohorts_evaluation_cohorts_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/CohortListItem"
         },
         "title": "Response List Cohorts Evaluation Cohorts Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/evaluation/cohorts/{cohort_id}": {
   "get": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica",
     "Cohortes"
    ],
    "summary": "Una cohorte congelada: definici\u00f3n, snapshot, resumen y filas",
    "description": "Devuelve la cohorte entera menos los bytes del archivo.\n\nSe publica su SHA-256 para poder cotejarla contra el archivo que uno tenga;\nentregar el contenido es otra operaci\u00f3n y no un efecto secundario de mirar.",
    "operationId": "get_cohort_evaluation_cohorts__cohort_id__get",
    "parameters": [
     {
      "name": "cohort_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Cohort Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CohortRecord"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/evaluation/preflight": {
   "post": {
    "tags": [
     "Evaluaci\u00f3n cient\u00edfica"
    ],
    "summary": "Comprobaci\u00f3n previa factual de una corrida (no ejecuta docking)",
    "description": "Inspecciona receptor, ligando y configuraci\u00f3n efectiva antes de ejecutar.\n\nNo valida el SMILES con `validate_smiles_or_raise`: un SMILES inv\u00e1lido no\npuede terminar en un 422 sin informe. Tiene que salir como un bloqueante\ndentro del propio preflight, que es donde el usuario lo va a leer.",
    "operationId": "evaluation_preflight_evaluation_preflight_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreflightRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Evaluation Preflight Evaluation Preflight Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/evaluation/batch": {
   "post": {
    "tags": [
     "Batch Screening"
    ],
    "summary": "Submit Batch",
    "description": "Virtual screening batch. Subi un archivo con SMILES, los evalua contra el target,\ny devuelve resultados rankeados.\n\nMulti-target: usa target_pdb_id=ALL para correr contra todos los targets pre-curados.\nEarly Exit: si esta activado, MolGraph pre-filtra moleculas similares a inactivos conocidos.\nLabel opcional: si el CSV/Excel tiene columna 'active' (1/0), se computa EF/AUC.",
    "operationId": "submit_batch_evaluation_batch_post",
    "parameters": [
     {
      "name": "target_pdb_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "minLength": 4,
       "maxLength": 10,
       "description": "Target PDB ID. Usar 'ALL' para correr contra todos los targets pre-curados.",
       "default": "7E2Y",
       "title": "Target Pdb Id"
      },
      "description": "Target PDB ID. Usar 'ALL' para correr contra todos los targets pre-curados."
     },
     {
      "name": "num_workers",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 6,
       "minimum": 1,
       "default": 2,
       "title": "Num Workers"
      }
     },
     {
      "name": "early_exit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "RETIRADO (BATCH-SCI-001). Se acepta por compatibilidad y se ignora: el pre-filtro de vecindad saltaba el acoplamiento y publicaba ceros como si fueran medidas. La respuesta declara `early_exit_enabled: false`.",
       "deprecated": true,
       "default": true,
       "title": "Early Exit"
      },
      "description": "RETIRADO (BATCH-SCI-001). Se acepta por compatibilidad y se ignora: el pre-filtro de vecindad saltaba el acoplamiento y publicaba ceros como si fueran medidas. La respuesta declara `early_exit_enabled: false`.",
      "deprecated": true
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_submit_batch_evaluation_batch_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Submit Batch Evaluation Batch Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/evaluation/batch/{batch_id}": {
   "get": {
    "tags": [
     "Batch Screening"
    ],
    "summary": "Get Batch Status",
    "description": "Get batch evaluation status and partial results.",
    "operationId": "get_batch_status_evaluation_batch__batch_id__get",
    "parameters": [
     {
      "name": "batch_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Batch Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Batch Status Evaluation Batch  Batch Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_cohort_preflight_evaluation_cohorts_preflight_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "description": "CSV, Excel, SDF o SMILES/TXT con las mol\u00e9culas."
     },
     "study": {
      "type": "string",
      "title": "Study",
      "description": "`CohortStudy` v1 serializado como JSON."
     }
    },
    "type": "object",
    "required": [
     "file",
     "study"
    ],
    "title": "Body_cohort_preflight_evaluation_cohorts_preflight_post"
   },
   "Body_create_cohort_evaluation_cohorts_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "description": "El MISMO archivo que se comprob\u00f3."
     },
     "study": {
      "type": "string",
      "title": "Study",
      "description": "El MISMO `CohortStudy` v1 que se comprob\u00f3."
     },
     "expected_fingerprint": {
      "type": "string",
      "title": "Expected Fingerprint",
      "description": "El `cohort_fingerprint` que devolvi\u00f3 la comprobaci\u00f3n previa."
     }
    },
    "type": "object",
    "required": [
     "file",
     "study",
     "expected_fingerprint"
    ],
    "title": "Body_create_cohort_evaluation_cohorts_post"
   },
   "Body_submit_batch_evaluation_batch_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_submit_batch_evaluation_batch_post"
   },
...
   "CohortPreflightResult": {
    "properties": {
     "schema_version": {
      "type": "integer",
      "const": 1,
      "title": "Schema Version",
      "default": 1
     },
     "generated_at": {
      "type": "string",
      "title": "Generated At"
     },
     "cohort_fingerprint": {
      "type": "string",
      "title": "Cohort Fingerprint"
     },
     "normalized_study": {
      "$ref": "#/components/schemas/CohortStudy"
     },
     "decision": {
      "type": "string",
      "enum": [
       "ready",
       "blocked"
      ],
      "title": "Decision"
     },
     "blockers": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Blockers"
     },
     "warnings": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Warnings"
     },
     "summary": {
      "$ref": "#/components/schemas/CohortSummary"
     },
     "rows": {
      "items": {
       "$ref": "#/components/schemas/CohortRow"
      },
      "type": "array",
      "title": "Rows"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "generated_at",
     "cohort_fingerprint",
     "normalized_study",
     "decision",
     "summary"
    ],
    "title": "CohortPreflightResult",
    "description": "Resultado de la comprobaci\u00f3n previa. No ejecuta nada y no guarda nada.\n\nSuperar esta comprobaci\u00f3n **no predice uni\u00f3n, actividad ni calidad\nfarmacol\u00f3gica**. Dice una cosa mucho m\u00e1s peque\u00f1a y mucho m\u00e1s comprobable:\nque estas entradas se pueden leer y que esta configuraci\u00f3n se puede\naplicar a todas por igual."
   },
...
   "CohortRecord": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "schema_version": {
      "type": "integer",
      "title": "Schema Version"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "cohort_fingerprint": {
      "type": "string",
      "title": "Cohort Fingerprint"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "source": {
      "$ref": "#/components/schemas/CohortSourceFile"
     },
     "provenance": {
      "$ref": "#/components/schemas/CohortProvenance"
     },
     "preflight": {
      "$ref": "#/components/schemas/CohortPreflightResult"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "schema_version",
     "name",
     "status",
     "cohort_fingerprint",
     "created_at",
     "source",
     "provenance",
     "preflight"
    ],
    "title": "CohortRecord",
    "description": "La cohorte completa: definici\u00f3n, snapshot, resumen y filas.\n\nLos cuatro est\u00e1n, y **una sola vez**: `preflight` es el dictamen congelado\nentero, as\u00ed que la definici\u00f3n es `preflight.normalized_study`, el resumen es\n`preflight.summary` y las filas son `preflight.rows`. Repetirlos en el nivel\nsuperior dar\u00eda tres copias del mismo dato que podr\u00edan divergir el d\u00eda que\nalguien edite una.\n\nLo que NO lleva son los bytes del archivo. Se guardan \u2014y su SHA-256 se\npublica aqu\u00ed para que se puedan cotejar\u2014, pero entregarlos es otra\noperaci\u00f3n, no un efecto secundario de mirar una cohorte."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}